+++ src/UI/SignTransactionsModals/types/signTransactionsModals.types.ts
@@ -20,11 +20,12 @@ export type ScreenType = (signProps: SignPropsType) => JSX.Element;
 export interface CustomConfirmScreensType {
   Ledger?: ScreenType;
   Extension?: ScreenType;
+  Passkey?: ScreenType;
   Metamask?: ScreenType;
   Opera?: ScreenType;
   CrossWindow?: ScreenType;
   WalletConnect?: ScreenType;
   Wallet?: ScreenType;
   Extra?: ScreenType;
-  MetamaskProxy?: ScreenType;
+  Iframe?: ScreenType;
 }