+++ src/UI/pages/UnlockPage/index.tsx
@@ -3,9 +3,10 @@ import classNames from 'classnames';
 import { withStyles, WithStylesImportType } from 'hocs/withStyles';
 import { useGetLoginInfo } from 'hooks';
 import { ExtensionLoginButton } from 'UI/extension/ExtensionLoginButton';
+import { IframeButton } from 'UI/iframe/IframeLoginButton';
 import { LedgerLoginButton } from 'UI/ledger/LedgerLoginButton';
-import { MetamaskProxyButton } from 'UI/metamaskProxy/MetamaskProxyLoginButton';
 import { OperaWalletLoginButton } from 'UI/operaWallet/OperaWalletLoginButton';
+import { PasskeyLoginButton } from 'UI/passkey/PasskeyLoginButton';
 import { WalletConnectLoginButton } from 'UI/walletConnect/WalletConnectLoginButton';
 import { WebWalletLoginButton } from 'UI/webWallet/WebWalletLoginButton';
 
@@ -16,9 +17,10 @@ export interface Props {
   loginRoute: string;
   LedgerLoginButtonText?: string;
   ExtensionLoginButtonText?: string;
+  PasskeyLoginButtonText?: string;
   OperaWalletLoginButtonText?: string;
   CrossWindowLoginButtonText?: string;
-  MetamaskProxyLoginButtonText?: string;
+  IframeLoginButtonText?: string;
   WebWalletLoginButtonText?: string;
   WalletConnectLoginButtonText?: string;
   WalletConnectV2LoginButtonText?: string;
@@ -33,8 +35,9 @@ const UnlockPageComponent = ({
   description = 'Pick a login method',
   WalletConnectLoginButtonText = 'xPortal App',
   ExtensionLoginButtonText = 'Extension',
+  PasskeyLoginButtonText = 'Passkey',
   OperaWalletLoginButtonText = 'Opera Crypto Wallet',
-  MetamaskProxyLoginButtonText = 'Metamask Proxy',
+  IframeLoginButtonText = 'Embeded web wallet',
   WebWalletLoginButtonText = 'Web wallet',
   globalStyles,
   styles
@@ -87,9 +90,14 @@ const UnlockPageComponent = ({
               loginButtonText={ExtensionLoginButtonText}
             />
 
-            <MetamaskProxyButton
+            <PasskeyLoginButton
               callbackRoute={loginRoute}
-              loginButtonText={MetamaskProxyLoginButtonText}
+              loginButtonText={PasskeyLoginButtonText}
+            />
+
+            <IframeButton
+              callbackRoute={loginRoute}
+              loginButtonText={IframeLoginButtonText}
             />
 
             <OperaWalletLoginButton